backend/requirements.txt
python-dotenv
pymongo
flask
flask-cors